+++ powerbox/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.5.3"
+__version__ = "0.5.4"
 
 from .powerbox import PowerBox, LogNormalPowerBox
 from .tools import get_power, angular_average